requirements.txt
python-dotenv==1.2.1
tmdbsimple==2.9.1
